auth bag and course info change ui and other bags

+++ lib/model/course/course_price/response/course_with_price_response.dart
@@ -36,7 +36,7 @@ class CourseWithPriceData with _$CourseWithPriceData {
   const factory CourseWithPriceData({
     @JsonKey(name: "course") required Course course,
     @JsonKey(name: "mentor")  Mentor? mentor,
-    @JsonKey(name: "price")  double? price,
+    @JsonKey(name: "price")  int? price,
     @JsonKey(name: "reception_counter")
     required ReceptionCounter? receptionCounter,
   }) = _CourseWithPriceData;
